pyproject.toml
[tool.poetry]
name = "flaskgeekbrains"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.3"
Flask-SQLAlchemy = "^3.0.3"
psycopg2 = "^2.9.6"
flask_login = "^0.6.2"
Flask-Login = "^0.6.2"
Flask-Migrate = "^4.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
